프로젝트 최종

+++ samjo/src/main/webapp/WEB-INF/views/board2/read.jsp
@@ -13,7 +13,7 @@ ArrayList<CmtDTO> list = new ArrayList<CmtDTO>();
 list = dao.cmtList(Integer.parseInt(request.getParameter("bbs_idx")));
 %>
 <jsp:useBean id="dto1" class="kr.co.samjo.board2.CmtDTO" scope="page"></jsp:useBean>
-<jsp:useBean id="dto2" class="kr.co.samjo.board2.boardDTO" scope="page"></jsp:useBean>
+<jsp:useBean id="dto3" class="kr.co.samjo.board2.boardDTO" scope="page"></jsp:useBean>
 <!-- 본문 시작 list.jsp -->
 <aside id="fh5co-hero-T">
 	<div class="flexslider">
@@ -21,26 +21,56 @@ list = dao.cmtList(Integer.parseInt(request.getParameter("bbs_idx")));
 			<div class="slider-text-inner desc">
 				<h2
 					style="margin-top: 300px; text-align: center; font-weight: bold;"
-					class="heading-section">${dto.bbs_title}</h2>
+					class="heading-section">자유게시판</h2>
 			</div>
 		</div>
 	</div>
-	<div class="row-read">
+	<div class="row-notice">
 		<div class="col-md-12 text-center project">
 			<div class="grid-project">
-				<ul>
-					<li>아이디 : ${dto.bbs_id}</li>
-					<li>제목 : ${dto.bbs_title}</li>
-				</ul>
-				<img src="../../storage/${dto.bbs_img}" class="img-responsive"
-					style="width: 40%; float: left;">
-				<div>${dto.bbs_content}</div>
+				<table class="table table-sm">
+					<thead>
+						<tr>
+							<th colspan="4" scope="col" style="font-size: 25px;">${dto.bbs_title}</th>
+						</tr>
+					</thead>
+					<tbody>
+						<tr>
+							<td colspan="2" style="text-align: left">아이디 : ${dto.bbs_id}</td>
+							<td>등록일 : ${dto.bbs_date}</td>
+							<td>조회수 : ${dto.bbs_count}</td>
+						</tr>
+						<tr>
+							<td colspan="4" rowspan="4" style="text-align: left">
+								<div>
+									<div style="font-size: 20px;">${dto.bbs_content}</div>
+									<c:if test="${dto.bbs_img != null}">
+										<img src="../../storage/${dto.bbs_img}">
+									</c:if>
+								</div>
+							</td>
+						</tr>
+					</tbody>
+				</table>
+
+
+
 			</div>
 			<div class='bottom'>
-				<input type="button" value="수정"
-					onclick="location.href='/board/update.do?bbs_idx=${dto.bbs_idx}'">
-				<input type="button" value="삭제"
-					onclick="location.href='/board/delete.do?bbs_idx=${dto.bbs_idx}'">
+				<%
+				if (s_mlevel.equals("2")) {
+				%>
+				<c:if test="${dto.bbs_id eq s_id }">
+					<div style="text-align: right;">
+						<input type="button" value="수정" class="btn btn-secondary"
+							onclick="location.href='/board/update.do?bbs_idx=${dto.bbs_idx}'">
+						<input type="button" value="삭제" class="btn btn-secondary"
+							onclick="location.href='/board/delete.do?bbs_idx=${dto.bbs_idx}'">
+					</div>
+				</c:if>
+				<%
+				}
+				%>
 				<button type="button" class="btn btn-secondary"
 					onclick="location.href='/board/List.do'">목록</button>
 			</div>
@@ -59,15 +89,22 @@ list = dao.cmtList(Integer.parseInt(request.getParameter("bbs_idx")));
 						<textarea class="form-control" id="cmt_content" name="cmt_content"
 							rows="3"></textarea>
 					</div>
+					<%
+					if (s_mlevel.equals("2")) {
+					%>
 					<button type="submit" class="btn btn-primary">등록</button>
+					<%
+					}
+					%>
+
 				</form>
-				<table class="table table-hover">
+				<table class="table table-hover cmt">
 					<thead>
 						<tr>
 							<th class="board_no">내용</th>
 							<th class="board_title">작성자</th>
 							<th class="board_date">작성일</th>
-							<th class="board_readcnt">답글</th>
+							<th class="board_readcnt"></th>
 						</tr>
 					</thead>
 					<thead>
@@ -80,19 +117,27 @@ list = dao.cmtList(Integer.parseInt(request.getParameter("bbs_idx")));
 							<td>
 								<%
 								for (int j = 0; j < dto.getCmt_re_setp(); j++) {
-								%> <img
-								src='../storage/reply.gif'> <%
+								%> <img src='../storage/reply.gif'> <%
  }
  %> <%=dto.getCmt_content()%>
 							</td>
 							<td><%=dto.getCmt_id()%></td>
 							<td><%=dto.getCmt_date()%></td>
-							<td><input type="button" value="답글"
+							<td>
+								<%
+								if (s_mlevel.equals("2")) {
+								%> <input type="button" value="답글"
 								onclick="location.href='/cmt/replyproc.do?cmt_idx=<%=dto.getCmt_idx()%>&bbs_idx=<%=dto.getCmt_bbs_idx()%>'">
-								<input type="button" value="수정"
-								onclick="location.href='/cmt/update.do?cmt_idx=<%=dto.getCmt_idx()%>&bbs_idx=<%=dto.getCmt_bbs_idx()%>'">
-								<input type="button" value="삭제"
-								onclick="location.href='/cmt/delete.do?cmt_idx=<%=dto.getCmt_idx()%>&bbs_idx=<%=dto.getCmt_bbs_idx()%>'">
+
+								<c:if test="${dto.bbs_id eq s_id }">
+									<input type="button" value="수정"
+										onclick="location.href='/cmt/update.do?cmt_idx=<%=dto.getCmt_idx()%>&bbs_idx=<%=dto.getCmt_bbs_idx()%>'">
+									<input type="button" value="삭제"
+										onclick="location.href='/cmt/delete.do?cmt_idx=<%=dto.getCmt_idx()%>&bbs_idx=<%=dto.getCmt_bbs_idx()%>'">
+								</c:if> <%
+ }
+ %>
+
 							</td>
 						</tr>
 						<%
